+++ PrincipalJogo/src/Ente/Entidades/Personagem.cpp
@@ -1,6 +1,6 @@
 #include "../../../include/Ente/Entidades/Personagem.h"
 
-Personagem::Personagem(Coordenada tam, Coordenada pos, bool est, int v, int d, ID id):
+Personagem::Personagem(Coordenada tam, Coordenada pos, bool est, float v, float d, ID id):
     Entidade(id,tam,pos,est),
     vida(v),
     dano(d),
@@ -13,11 +13,11 @@ Personagem::~Personagem(){
 }
 
 
-void Personagem::setDano(int d){
+void Personagem::setDano(float d){
     dano = d;
 }
 
-void Personagem::receberDano(int d){
+void Personagem::receberDano(float d){
     if(vida > 0)
         vida -= d;
     else {
